start of work on the service selector role.
this lets you make a meta service that doesn't decide what actual
service it'll dispatch to until it's read the HTTP request in.

this paves the way for:

  -- name-based virtual hosts (a work-in-progress plugin is included
     with this commit)

  -- complex selection plugins that contact external resources to
     decide where to map to

this patch also rearranges a lot of the plug-in configuration
management commands to clean it up.

currently Vhosts.pm is hard-coded with some values and isn't usable
without some tweaking.

also, PUTs/POSTs are expected to still fail on some service types.


git-svn-id: http://code.sixapart.com/svn/perlbal/trunk@320 6caf28e9-730f-0410-b62b-a31386fe13fb