add tests for verify_backend,
and fix a bug that was uncovered by yesterday's assertion of parameter
validation assertions in HTTPHeaders {res/req}_keep_alive.  the bug
was that after an OPTIONS request, the BackendHTTP's req_headers was
still empty (since we never formally turned the OPTIONS * string into
a request object, just sent it) and the updated {res/req}_keep_alive
correctly expected non-undef values.  so now i add a new flag to say
that it's an options response.


git-svn-id: http://code.sixapart.com/svn/perlbal/trunk@391 6caf28e9-730f-0410-b62b-a31386fe13fb