Add 'DEFAULT' command
This command allows you to set defaults on the service objects you create.  Useful if you have many services and you set the same parameters on each of them.

Usage:

    DEFAULT persist_client  = true
    DEFAULT persist_backend = true

As with the rest of Perlbal configuration, this functions top down.  The defaults only apply to services created after the default is set.