reject PUTs via Content-MD5 header on corrupt transfers
If a client knows the MD5 checksum of what it is uploading ahead
of time, it may specify the Content-MD5 header and ask Perlbal
to verify it while the file is being read off the socket and
detect corrupted transfers before completion.

This allows to one to avoid rereading the entire file to verify
the MD5 post-upload.  I find this functionality useful because
TCP checksums are weak and I've experienced network corruption
that TCP did not detect several times over the years.

I hope I've implemented this correctly according to section
14.15 of RFC 2616.  I don't know if there are other examples of
other HTTP servers that can reject PUTs based on Content-MD5
(other servers not written by me, I don't trust myself with
this stuff).

Signed-off-by: Eric Wong <normalperson@yhbt.net>